+++ src/js/components/BookmarkForm.svelte
@@ -1,55 +1,126 @@
 <script>
-    import { createBookmark } from '../api/bookmarks'
-    import { t } from '../lib/translate'
+    import browser from 'webextension-polyfill'
+    import { equals, pick } from 'rambda'
+    import { onMount } from 'svelte'
+
+    import { messages } from '../lib/messaging'
+    import { humanizeDate, t } from '../lib/translate'
+    import { dropdownShown } from '../stores/dropdown'
+    import { search } from '../stores/search'
+
+    import Button from './Button.svelte'
+    import CategorySearch from './CategorySearch.svelte'
     import CategorySelector from './CategorySelector.svelte'
+    import ChildCategories from './ChildCategories.svelte'
+    import Dropdown from './Dropdown.svelte'
+    import DropdownGroup from './DropdownGroup.svelte'
+    import DropdownToggles from './DropdownToggles.svelte'
     import Favicon from './Favicon.svelte'
+    import Icon from './Icon.svelte'
+    import IconFa from './IconFa.svelte'
     import InputGroup from './form/InputGroup.svelte'
+    import RecentCategories from './RecentCategories.svelte'
+    import SearchResults from './SearchResults.svelte'
 
     export let bookmark
     let form
-    let submitButton
 
-    const onSubmit = async (event) => {
-        if (!form.reportValidity()) {
-            return false
-        }
+    const isVisible = (dropdown) => equals($dropdownShown, dropdown)
 
-        const params = {
-            parentId: form.category.value,
-            title: form.title.value,
-            url: form.url.value,
-        }
-        const subcategory = form.subcategory && form.subcategory.value
-        // console.debug('[BookmarkForm] Submitted:', params, subcategory)
+    function onSearchFocus () {
+        // console.debug('[CategorySelector] onSearchFocus')
+        $dropdownShown = 'search'
+    }
 
-        if (subcategory) {
-            const newSubcategory = await createBookmark({
-                parentId: params.parentId,
-                title: subcategory,
+    function updateCategories (results) {
+        console.debug('[CategorySelector] updateCategories:', results.length)
+        $search.results = results
+        $search.last = $search.query
+        $dropdownShown = 'search'
+    }
+
+    async function onSubmit (event) {
+        if (!form.reportValidity()) return false
+
+        const data = Object.fromEntries(new FormData(form))
+        const bookmarkFields = ['parentId', 'title', 'url']
+        // console.debug('[BookmarkForm] Submitted:', data, data.subcategory)
+
+        if (data.subcategory) {
+            const newSubcategory = await browser.bookmarks.create({
+                parentId: data.parentId,
+                title: data.subcategory,
             })
-
-            params.parentId = newSubcategory.id
+            data.parentId = newSubcategory.id
         }
 
-        const newBookmark = await createBookmark(params)
+        const newBookmark = await browser.bookmarks.create(pick(bookmarkFields, data))
         console.info('[BookmarkForm] Bookmark saved:', newBookmark)
 
-        window.close()
+        location.reload() // Refresh the popup
         return false
     }
+
+    onMount(() => {
+        messages.on('categorySearch', updateCategories)
+    })
 </script>
 
-<style>
-    .btn[type=submit] {
-        padding: .25rem .8rem;
-    }
-</style>
-
-<form id="bookmarkForm"
+<form class="bookmark-form"
       bind:this={ form }
       on:submit|preventDefault={ onSubmit }
       >
-    <CategorySelector />
+    {#if bookmark.dateAdded }
+    <p>{ t('added') }: { humanizeDate(bookmark.dateAdded) }</p>
+    {/if}
+    {#if bookmark.parentId }
+    <p>{ t('category') }: { bookmark.parentId }</p>
+    {/if}
+
+    <div class="form-group">
+        <CategorySelector>
+            <span slot="status">
+                {#if $search.last }
+                    <span class="label" title="{ t('search') }">
+                        <IconFa icon="search" />
+                        { $search.last }
+                    </span>
+                {/if}
+            </span>
+
+            <CategorySearch on:focus={ onSearchFocus } />
+
+            <DropdownToggles />
+        </CategorySelector>
+
+        <DropdownGroup>
+            <Dropdown name={'search'}>
+                {#if isVisible('search') && $search.last }
+                    <SearchResults categories={$search.results} />
+                {/if}
+            </Dropdown>
+
+            <Dropdown name={'children'}>
+                <ChildCategories />
+            </Dropdown>
+
+            <Dropdown name={'recent'}>
+                <RecentCategories />
+            </Dropdown>
+        </DropdownGroup>
+    </div>
+
+    <div class="form-group subcategory"
+         class:active={ $dropdownShown === 'subcategory' }
+         class:d-hide={ $dropdownShown !== 'subcategory' }
+         >
+        <InputGroup name="subcategory"
+                    label={ t('add_subcategory') }>
+            <Button name="toggleSubcategory" classes="input-group-btn">
+                <Icon icon="cross" />
+            </Button>
+        </InputGroup>
+    </div>
 
     <div class="form-group">
         <InputGroup name="url"
@@ -68,7 +139,6 @@
 
     <div class="form-group text-right buttons-row">
         <button type="submit"
-                bind:this={submitButton}
                 class="btn btn-primary">
             { t('buttons_add') }
         </button>