+++ src/js/components/CategorySearch.svelte
@@ -1,26 +1,29 @@
 <script>
     import { length } from 'rambda'
+    import { onMount } from 'svelte'
+
     import { inputEvent$ } from '../lib/reactive'
     import { messages } from '../lib/messaging'
-    import { onMount } from 'svelte'
     import { t } from '../lib/translate'
+    import { search } from '../stores/search'
 
-    export let name
-    export let value = ''
-
-    let search = ''
+    let input = ''
 
     function searchCategories (query) {
-        messages.emit('api', { type: 'categorySearch', query })
+        // console.debug('[CategorySearch] search:', query)
+        messages.emit('api', { action: 'categorySearch', query })
     }
 
     function clearSearch () {
-        // value = ''
+        // console.debug('[CategorySearch] clearSearch')
+        // $search.query = ''
+        $search.last = ''
         messages.emit('search:clear')
-        search.focus()
+        input.focus()
     }
 
     function onInput (query) {
+        // console.debug('[CategorySearch] onInput:', query)
         if (length(query) >= 2) {
             // console.debug('[CategorySearch] >>> Will search with query:', query)
             searchCategories(query)
@@ -32,22 +35,21 @@
 
     onMount(() => {
         const options = { minLength: 0, debounceTime: 400 }
-        const categorySearch$ = inputEvent$(search, options)
+        const categorySearch$ = inputEvent$(input, options)
 
         categorySearch$
             .throttle(1000, { leading: true, trailing: false })
             // .spy('Throttled:')
             .observe(onInput, console.error)
 
-        search.focus()
+        input.focus()
     })
 </script>
 
-<svelte:options accessors />
-<input {name} required
+<input name="search"
        class="form-input"
-       bind:this={search}
-       bind:value={value}
+       bind:this={ input }
+       bind:value={ $search.query }
        on:focus
        placeholder={ t('search_placeholder') }
        autocomplete="off">