+++ src/js/components/ChildCategories.svelte
@@ -1,17 +1,19 @@
 <script>
+    import browser from 'webextension-polyfill'
     import { filter } from 'rambda'
     import { onDestroy, onMount } from 'svelte'
-    import { getChildren } from '../api/categories'
+
     import { isCategory } from '../api/helpers'
     import { messages } from '../lib/messaging'
+    import { t } from '../lib/translate'
     import { sortByTitleCaseInsensitive } from '../lib/pure'
-    import CategoryList from './CategoryList.svelte'
+    import CategoryMenu from './CategoryMenu.svelte'
 
     export let children = []
 
     async function updateChildren (category) {
         // console.debug('[ChildCategories] updateChildren:', category)
-        const results = await getChildren(category.id)
+        const results = await browser.bookmarks.getChildren(category.id)
         children = sortByTitleCaseInsensitive(
             filter(isCategory, results)
         )
@@ -27,7 +29,10 @@
 </script>
 
 <ul class="menu" tabindex="-1"
-    on:categorySelection={updateChildren}
-    >
-    <CategoryList categories={children} />
+    on:categorySelection={updateChildren}>
+    <small class="toast">
+        { t('subcategories') }
+        ({ children.length } { t('pieces') })
+    </small>
+    <CategoryMenu categories={children} />
 </ul>