+++ src/js/components/Popup.svelte
@@ -1,47 +1,77 @@
 <script>
+    import browser from 'webextension-polyfill'
+    import { indexBy, prop, sortBy, toPairs } from 'rambda'
     import { onDestroy, onMount } from 'svelte'
 
-    import { messages } from '../lib/messaging'
-    import { currentTab } from '../stores/currentTab'
     import { getCurrentTab } from '../api/tabs'
+    import { messages } from '../lib/messaging'
     import { t } from '../lib/translate'
+
     import BookmarkForm from '../components/BookmarkForm.svelte'
     import CloseButton from '../components/CloseButton.svelte'
     import Explore from '../components/Explore.svelte'
 
-    let savedBookmarks = []
+    import { currentTab } from '../stores/currentTab'
+    import { savedBookmarks } from '../stores/savedBookmarks'
 
-    $: bookmarkCount = savedBookmarks.length
+    import Bookmark from './Bookmark.svelte'
+    import BookmarkForm from './BookmarkForm.svelte'
+    import CloseButton from './CloseButton.svelte'
+
+    $: bookmarkCount = $savedBookmarks.size
     $: popupHeader = (
         bookmarkCount >= 1
             ? t('saved_bookmark')
             : t('add_bookmark')
     )
 
+    async function deleteBookmark (bookmark) {
+        // console.debug('Deleting bookmark:', bookmark)
+        browser.bookmarks.remove(bookmark.id)
+            .then(() => {
+                // console.info('Bookmark deleted:', bookmark)
+                location.reload()
+            })
+            .then(updateStatus)
+            .catch(console.error)
+    }
+
     function onClose () {
         window.close()
         return false
     }
 
+    function updateStatus () {
+        getCurrentTab().then(tab => {
+            // console.debug('[Popup] updating bookmark status:', { tab })
+            $currentTab = { ...tab }
+            messages.emit('api', { action: 'bookmarkStatus', tab })
+        })
+    }
+
     function updateBookmarks (bookmarks) {
-        // console.log('[Popup] updateBookmarks:', bookmarks)
-        savedBookmarks = bookmarks
+        const sorted = sortBy(prop('dateAdded'), bookmarks || [])
+        const saved = new Map(toPairs(indexBy(prop('id'), sorted)))
+        // console.debug('[Popup] updateBookmarks sorted:', { sorted, saved })
+
+        // TODO Use separate store?
+        $savedBookmarks = saved
     }
 
     onMount(() => {
         messages.on('bookmarkStatus', updateBookmarks)
         messages.on('button:close', onClose)
+        messages.on('deleteBookmark', deleteBookmark)
 
-        getCurrentTab().then(tab => {
-            // console.debug('[Popup] current tab:', tab)
-            $currentTab = { ...tab }
-            messages.emit('api', { type: 'bookmarkStatus', tab })
-        })
+        messages.emit('api', { action: 'recentCategories' })
+
+        updateStatus()
     })
 
     onDestroy(() => {
         messages.off('bookmarkStatus', updateBookmarks)
         messages.off('button:close', onClose)
+        messages.off('deleteBookmark', deleteBookmark)
     })
 </script>
 
@@ -56,6 +86,11 @@
         </h1>
     </div>
     <div class="card-body">
-        <BookmarkForm bookmark={$currentTab} />
+        <div class="saved-bookmarks">
+            {#each [...$savedBookmarks.values()] as bookmark (bookmark.id)}
+                <Bookmark {bookmark} />
+            {/each}
+        </div>
+        <BookmarkForm bookmark={ $currentTab } />
     </div>
 </div>