+++ src/js/components/RecentCategories.svelte
@@ -1,7 +1,9 @@
 <script>
-    import { messages } from '../lib/messaging'
     import { onDestroy, onMount } from 'svelte'
-    import CategoryList from './CategoryList.svelte'
+
+    import { messages } from '../lib/messaging'
+    import { t } from '../lib/translate'
+    import CategoryMenu from './CategoryMenu.svelte'
 
     export let recentCategories = []
 
@@ -20,5 +22,6 @@
 </script>
 
 <ul class="menu" tabindex="-1">
-    <CategoryList categories={recentCategories} />
+    <small class="toast">{ t('recent_categories') }</small>
+    <CategoryMenu categories={recentCategories} />
 </ul>