+++ src/js/components/SearchResults.svelte
@@ -1,18 +1,20 @@
 <script>
     import { isEmpty, not } from 'rambda'
     import { t } from '../lib/translate'
-    import CategoryList from './CategoryList.svelte'
+    import CategoryMenu from './CategoryMenu.svelte'
     import MainCategories from './MainCategories.svelte'
 
     export let categories = []
 </script>
 
-{#if not(isEmpty(categories)) }
-    <ul class="menu" tabindex="-1">
-        <CategoryList categories={categories} />
+<ul class="menu" tabindex="-1">
+    <small class="toast">
+        { t('search_results') }
+        ({ categories.length } { t('results') })
+    </small>
+    {#if not(isEmpty(categories)) }
+        <CategoryMenu categories={categories} />
         <li class="divider" data-content={ t('root_categories') }></li>
         <MainCategories />
-    </ul>
-{:else}
-    <small>No categories found</small>
-{/if}
+    {/if}
+</ul>